requirements.txt
asgiref==3.3.4
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-cors-headers==3.7.0
django-import-export==2.8.0
django-jet-reboot==1.1.1
django-tinymce==3.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
fonttools==4.33.3
google-analytics==0.0.0
google-analytics-data==0.12.0
google-api-core==2.8.1
google-api-python-client==1.4.1
google-auth==2.6.6
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.19.1
idna==3.3
inflection==0.5.1
iso8601==1.0.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.5.2
more-itertools==8.13.0
numpy==1.22.4
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==8.2.0
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
requests==2.27.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sob==1.47.0
sqlparse==0.4.1
tablib==3.2.1
uritemplate==3.0.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
